requirements_actual.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
anyio==4.9.0
attrs==25.3.0
certifi==2025.7.14
cffi==1.17.1
click==8.2.1
colorama==0.4.6
comtypes==1.4.11
coverage==7.9.2
cryptography==45.0.5
edge-tts==7.0.2
frozenlist==1.7.0
idna==3.10
iniconfig==2.1.0
multidict==6.6.3
packaging==25.0
pluggy==1.6.0
propcache==0.3.2
psutil==7.0.0
pycparser==2.22
pygame==2.6.1
Pygments==2.19.2
pypiwin32==223
pytest==8.4.1
pytest-asyncio==1.1.0
pytest-cov==6.2.1
pytest-mock==3.14.1
pyttsx3==2.99
pywin32==311
PyYAML==6.0.2
sniffio==1.3.1
srt==3.5.3
tabulate==0.9.0
typing_extensions==4.14.1
-e git+https://github.com/KarlBotha/VPA.git@1c512eea2dbec9adf16e495092e2c45f4f0068fd#egg=vpa
yarl==1.20.1
